# Project-1.0/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type B2ObjectBody = string;

export interface B2UploadResult {
  storageKey: string;
  bucketId: string;
  /** @minimum 0 */
  sizeBytes: number;
  mimeType: string;
}

export interface B2SignedUrlResponse {
  signedUrl: string;
  /**
     * @minimum 60
     * @maximum 3600
     */
  expiresIn: number;
}

export type B2VideoInspectResponseCodec = {
  found?: string;
  [key: string]: unknown;
};

export type B2VideoInspectResponseAtoms = { [key: string]: unknown };

export interface B2VideoInspectResponse {
  fileId: string;
  fileName: string;
  mimeType: string;
  /** @minimum 0 */
  size: number;
  codec: B2VideoInspectResponseCodec;
  moovBeforeMdat: boolean;
  moovFound: boolean;
  mdatFound: boolean;
  atoms: B2VideoInspectResponseAtoms;
  /** @minimum 0 */
  headerBytesRead: number;
}

export type GetB2ObjectParams = {
download?: string;
};
export type GetB2SignedUrlParams = {
/**
 * @minimum 60
 * @maximum 3600
 */
expiresIn?: number;
download?: string;
};