# Project-1.0/lib/api-zod/src/generated/types/b2SignedUrlResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface B2SignedUrlResponse {
  signedUrl: string;
  /**
     * @minimum 60
     * @maximum 3600
     */
  expiresIn: number;
}