# Project-1.0/lib/api-zod/src/generated/types/b2UploadResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface B2UploadResult {
  storageKey: string;
  bucketId: string;
  /** @minimum 0 */
  sizeBytes: number;
  mimeType: string;
}