# Project-1.0/lib/api-zod/src/generated/types/b2VideoInspectResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { B2VideoInspectResponseAtoms } from './b2VideoInspectResponseAtoms';
import type { B2VideoInspectResponseCodec } from './b2VideoInspectResponseCodec';

export interface B2VideoInspectResponse {
  fileId: string;
  fileName: string;
  mimeType: string;
  /** @minimum 0 */
  size: number;
  codec: B2VideoInspectResponseCodec;
  moovBeforeMdat: boolean;
  moovFound: boolean;
  mdatFound: boolean;
  atoms: B2VideoInspectResponseAtoms;
  /** @minimum 0 */
  headerBytesRead: number;
}